+++ src/constants.js
@@ -1,3 +1,4 @@
+const DEBOUNCE_VALUE_DEFAULT = 0;
 const POSITION_PROP_DEFAULT = 'position';
 const RENDER_ON_RESIZE_DEFAULT = true;
 const SIZE_PROP_DEFAULT = 'size';
@@ -39,6 +40,26 @@ const DOM_ELEMENT_SIZE_KEYS = [
   'scrollWidth'
 ];
 
+const NATURAL_REGEXP = /natural/;
+
+const VOID_ELEMENT_TAG_NAMES = [
+  'AREA',
+  'BASE',
+  'BR',
+  'COL',
+  'EMBED',
+  'HR',
+  'IMG',
+  'INPUT',
+  'LINK',
+  'MENUITEM',
+  'META',
+  'PARAM',
+  'SOURCE',
+  'TRACK',
+  'WBR'
+];
+
 const ALL_DOM_ELEMENT_KEYS = [
   ...DOM_ELEMENT_POSITION_KEYS,
   ...DOM_ELEMENT_SIZE_KEYS
@@ -59,16 +80,6 @@ const ALL_KEYS = [
   ...ALL_SIZE_KEYS
 ];
 
-let initialState = {},
-    index = ALL_KEYS.length,
-    key;
-
-while (index--) {
-  key = ALL_KEYS[index];
-
-  initialState[key] = 0;
-}
-
 export {ALL_BOUNDING_CLIENT_RECT_KEYS};
 export {ALL_DOM_ELEMENT_KEYS};
 export {ALL_KEYS};
@@ -76,9 +87,11 @@ export {ALL_POSITION_KEYS};
 export {ALL_SIZE_KEYS};
 export {BOUNDING_CLIENT_RECT_POSITION_KEYS};
 export {BOUNDING_CLIENT_RECT_SIZE_KEYS};
+export {DEBOUNCE_VALUE_DEFAULT};
 export {DOM_ELEMENT_POSITION_KEYS};
 export {DOM_ELEMENT_SIZE_KEYS};
-export {initialState};
+export {NATURAL_REGEXP};
 export {POSITION_PROP_DEFAULT};
 export {RENDER_ON_RESIZE_DEFAULT};
 export {SIZE_PROP_DEFAULT};
+export {VOID_ELEMENT_TAG_NAMES};