+++ src/utils.js
@@ -1,100 +1,14 @@
+// external dependencies
+import includes from 'lodash/includes';
+import isArray from 'lodash/isArray';
+import isUndefined from 'lodash/isUndefined';
+
 // constants
 import {
   ALL_POSITION_KEYS,
-  ALL_SIZE_KEYS
+  ALL_SIZE_KEYS,
+  NATURAL_REGEXP
 } from './constants';
-
-const NATURAL_REGEXP = /natural/;
-
-/**
- * get the toString value for the object
- *
- * @param {any} object
- * @returns {string}
- */
-const toString = (object) => {
-  return Object.prototype.toString.call(object);
-};
-
-/**
- * loop over array, executing fn
- *
- * @param {array<any>} array
- * @param {function} fn
- */
-const forEach = (array, fn) => {
-  const length = array.length;
-
-  let index = -1;
-
-  while (++index < length) {
-    fn(array[index], index, array);
-  }
-};
-
-/**
- * determine if object is an array
- *
- * @param {any} object
- * @returns {boolean}
- */
-const isArray = (object) => {
-  return toString(object) === '[object Array]';
-};
-
-/**
- * determine if object is an object
- *
- * @param {any} object
- * @returns {boolean}
- */
-const isObject = (object) => {
-  return toString(object) === '[object Object]' && !!object;
-};
-
-/**
- * determine if object is a string
- *
- * @param {any} object
- * @returns {boolean}
- */
-const isString = (object) => {
-  return toString(object) === '[object String]';
-};
-
-/**
- * determine if object is undefined
- *
- * @param {any} object
- * @returns {boolean}
- */
-const isUndefined = (object) => {
-  return object === void 0;
-};
-
-/**
- * determine if array contains item at one of the indices
- *
- * @param {array<any>} array
- * @param {any} item
- * @returns {boolean}
- */
-const arrayContains = (array, item) => {
-  return isArray(array) && !!~array.indexOf(item);
-};
-
-/**
- * get subset of array1 based on items existing in array2
- *
- * @param {array<*>} array1
- * @param {array<*>} array2
- * @returns {array<T>}
- */
-const arraySubset = (array1, array2) => {
-  return array1.filter((item) => {
-    return array2.includes(item);
-  });
-};
 
 /**
  * For naturalHeight and naturalWidth, coalesce the values
@@ -117,11 +31,11 @@ const getNaturalDimensionValue = (source, key) => {
  * create an object based on the keys passed and their value
  * in the source object
  *
- * @param {array<string>} keys
+ * @param {Array<string>} keys
  * @param {function} keys.reduce
- * @param {object|ClientRect} source
+ * @param {Object|ClientRect} source
  * @param {boolean} shouldAlterNaturalKeys=true
- * @returns {object}
+ * @returns {Object}
  */
 const createObjectFromKeys = (keys, source, shouldAlterNaturalKeys = true) => {
   return keys.reduce((target, key) => {
@@ -133,32 +47,56 @@ const createObjectFromKeys = (keys, source, shouldAlterNaturalKeys = true) => {
 };
 
 /**
+ * get subset of array1 based on items existing in array2
+ *
+ * @param {Array<*>} array1
+ * @param {Array<*>} array2
+ * @returns {Array<T>}
+ */
+const getArraySubset = (array1, array2) => {
+  return array1.filter((item) => {
+    return includes(array2, item);
+  });
+};
+
+/**
+ * wait to assign the raf until mount, so it has access to the
+ * window object
+ *
+ * @returns {function}
+ */
+const getRequestAnimationFrame = () => {
+  return (
+    window.requestAnimationFrame ||
+    window.webkitRequestAnimationFrame ||
+    window.mozRequestAnimationFrame ||
+    function(callback) {
+      window.setTimeout(callback, 1000 / 60);
+    }
+  );
+};
+
+/**
  * based on their existence in keysToTestAgainst, determine which of the keys
  * passed are considered valid
  *
- * @param {array<string>} keys
- * @param {array<string>} keysToTestAgainst
- * @returns {array<string>}
+ * @param {Array<string>} keys
+ * @param {Array<string>} keysToTestAgainst
+ * @returns {Array<string>}
  */
 const getValidKeys = (keys, keysToTestAgainst) => {
-  let validKeys = [];
-
-  forEach(keys, (key) => {
-    if (arrayContains(keysToTestAgainst, key)) {
-      validKeys.push(key);
-    }
+  return keys.filter((key) => {
+    return includes(keysToTestAgainst, key);
   });
-
-  return validKeys;
 };
 
 /**
  * get the position and size, and booleans to identify they're
  * intended existence in state
  *
- * @param {array<string>} keys
- * @param {object} currentState
- * @returns {{hasPosition: boolean, hasSize: boolean, position: object, size: object}}
+ * @param {Array<string>} keys
+ * @param {Object} currentState
+ * @returns {{hasPosition: boolean, hasSize: boolean, position: Object, size: Object}}
  */
 const getValuesProperties = (keys, currentState) => {
   if (isArray(keys)) {
@@ -167,11 +105,11 @@ const getValuesProperties = (keys, currentState) => {
         hasPosition = false,
         hasSize = false;
 
-    forEach(keys, (key) => {
-      if (arrayContains(ALL_POSITION_KEYS, key)) {
+    keys.forEach((key) => {
+      if (includes(ALL_POSITION_KEYS, key)) {
         position[key] = currentState[key];
         hasPosition = true;
-      } else if (arrayContains(ALL_SIZE_KEYS, key)) {
+      } else if (includes(ALL_SIZE_KEYS, key)) {
         size[key] = currentState[key];
         hasSize = true;
       }
@@ -201,11 +139,11 @@ const getValuesProperties = (keys, currentState) => {
  * or size or both properties that are objects containing the respective
  * values for the associated keys
  *
- * @param {array<string>} keys
- * @param {object} currentState
+ * @param {Array<string>} keys
+ * @param {Object} currentState
  * @param {string} positionProp
  * @param {string} sizeProp
- * @returns {object}
+ * @returns {Object}
  */
 const getValues = (keys, currentState, {positionProp, sizeProp}) => {
   const {
@@ -232,35 +170,39 @@ const getValues = (keys, currentState, {positionProp, sizeProp}) => {
  * iterate through keys and determine if the values have
  * changed compared to what is stored in state
  *
- * @param {array<string>} keys
- * @param {object} values
- * @param {object} currentState
+ * @param {Array<string>} keys
+ * @param {Object} values
+ * @param {Object} currentState
  * @returns {boolean}
  */
 const haveValuesChanged = (keys, values, currentState) => {
-  let index = -1,
-      key;
-
-  while (++index < keys.length) {
-    key = keys[index];
-
-    if (values[key] !== currentState[key]) {
-      return true;
-    }
-  }
-
-  return false;
+  return keys.some((key) => {
+    return values[key] !== currentState[key];
+  });
 };
 
-export {arrayContains};
-export {arraySubset};
+/**
+ * based on desiredKeys, build the initialState object
+ *
+ * @param {Array<string>} allKeys
+ * @param {Array<string>} desiredKeys
+ * @returns {Array<T>}
+ */
+const reduceStateToMatchingKeys = (allKeys, desiredKeys) => {
+  return allKeys.reduce((accumulatedInitialState, key) => {
+    if (includes(desiredKeys, key)) {
+      accumulatedInitialState[key] = 0;
+    }
+
+    return accumulatedInitialState;
+  }, {});
+};
+
+export {getArraySubset};
 export {createObjectFromKeys};
-export {forEach};
 export {getNaturalDimensionValue};
+export {getRequestAnimationFrame};
 export {getValidKeys};
 export {getValues};
 export {haveValuesChanged};
-export {isArray};
-export {isObject};
-export {isString};
-export {isUndefined};
+export {reduceStateToMatchingKeys};